Unit tests for Ufront/Hello
This is my first time trying to create a unit test suite for a Ufront app.

It seems to work, though I'm still trying to work on creating helpers / testing
APIs that are simple to understand and simple to write, so that testing is not
too cumbersome, especially with all the async and injection going on.

So I may restructure the way these work as I look for a cleaner way to write these tests.

But on the plus side: the basics work!

I've used Buddy, which is quite a nice BDD testing library.
Using Utest should also work well.

Please note I need to update some of the libs (particularly ufront-mvc) with
some changes I made in the course of preparing this.